+++ src/lazy/binary/encoded_value.rs
@@ -6,7 +6,7 @@ use std::ops::Range;
 /// Represents the type, offset, and length metadata of the various components of an encoded value
 /// in an input stream.
 ///
-/// Each [`LazyRawValue`](super::raw::lazy_raw_value::LazyRawBinaryValue) contains an `EncodedValue`,
+/// Each [`LazyRawValue`](super::raw::value::LazyRawBinaryValue) contains an `EncodedValue`,
 /// allowing a user to re-read (that is: parse) the body of the value as many times as necessary
 /// without re-parsing its header information each time.
 #[derive(Clone, Copy, Debug, PartialEq)]
