+++ src/lazy/binary/mod.rs
@@ -1,8 +1,7 @@
 mod encoded_value;
 pub mod immutable_buffer;
-pub mod lazy_reader;
 pub mod raw;
-pub mod system;
 
+pub(crate) mod format;
 #[cfg(test)]
 pub(crate) mod test_utilities;
