+++ src/lazy/binary/raw/annotations_iterator.rs
