--- src/lazy/binary/system/mod.rs
@@ -1,4 +1,0 @@
-pub mod lazy_sequence;
-pub mod lazy_struct;
-pub mod lazy_system_reader;
-pub mod lazy_value;
