+++ src/lazy/mod.rs
@@ -5,5 +5,10 @@ pub mod binary;
 pub mod format;
 pub mod raw_stream_item;
 pub mod raw_value_ref;
+pub mod reader;
+pub mod sequence;
+pub mod r#struct;
+pub mod system_reader;
 pub mod system_stream_item;
+pub mod value;
 pub mod value_ref;
